+++ frontend/src/components/annotation/AnnotationImage.tsx
@@ -2,30 +2,23 @@ import { useTranslation } from "react-i18next";
 import { capitalize, Grid } from "@mui/material";
 import AnnotationImageDisplay from "./AnnotationImageDisplay";
 import AnnotationImageNavigation from "./AnnotationImageNavigation";
+import { useMainContext } from "../../contexts/mainContext";
 
 
-export default function AnnotationImage (
-    props
-) {
+export default function AnnotationImage () {
     const { t } = useTranslation();
+    const { image } = useMainContext();
 
     return (
         <>
-            { props.image ? (
+            { image() ? (
                 <Grid 
                     container 
                     direction="column"
                     className="pageContainer"
                 >
-                    <AnnotationImageDisplay 
-                        image = { props.image }
-                        isAnnotated = { props.isAnnotated }
-                    />
-                    <AnnotationImageNavigation
-                        previous={ props.previous }
-                        next={ props.next }
-                        lastOrFirstImage={ props.lastOrFirstImage }
-                    />
+                    <AnnotationImageDisplay />
+                    <AnnotationImageNavigation />
                 </Grid>
             ) : (
                 <p>
