+++ frontend/src/components/annotation/AnnotationSaveError.tsx
@@ -1,11 +1,15 @@
 import { Dialog, DialogActions, DialogContent, DialogTitle, Typography, capitalize } from "@mui/material";
 import { useTranslation } from "react-i18next";
+import { useAnnotationContext } from "../../contexts/annotationContext";
 import ButtonCancel from "../common/buttonCancel";
 
-const AnnotationSaveError = (props) => {
-    const { t } = useTranslation()
+const AnnotationSaveError = () => {
+    const { t } = useTranslation();
+    const { openSaveErrorDialog, handleCloseSaveErrorDialog } = useAnnotationContext();
+
+
     return (
-        <Dialog open={props.openSaveErrorDialog}>
+        <Dialog open={openSaveErrorDialog}>
             <DialogTitle>
                 { capitalize(t("annotations.cannot_save")) }
             </DialogTitle>
@@ -17,7 +21,10 @@ const AnnotationSaveError = (props) => {
             </DialogContent>
 
             <DialogActions>
-                <ButtonCancel content={ capitalize(t("annotations.fix_my_data")) } cancel={ props.handleCloseSaveErrorDialog } />
+                <ButtonCancel 
+                    content={ capitalize(t("annotations.fix_my_data")) } 
+                    cancel={ handleCloseSaveErrorDialog } 
+                />
             </DialogActions>
         </Dialog>
     )
