+++ frontend/src/components/annotation/TraitInput.tsx
@@ -10,7 +10,6 @@ const traitList = {
 };
 
 
-
 const TraitInput = (
     props
 ) => {
